edited 5-sqrt_recursion.c

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -11,13 +11,13 @@
 int find_sqrt(int n, int start, int end)
 {
 	int mid;
-	int square;
+	long square;
 
 	if (start > end)
 		return (-1);
 
 	mid = start + (end - start) / 2;
-	square = mid * mid;
+	square = (long) mid * mid;
 
 	if (square == n)
 		return (mid);
